+++ pom.xml
@@ -38,6 +38,7 @@
             <version>5.9.1</version>
             <scope>test</scope>
         </dependency>
+
     </dependencies>
 
     <build>
