+++ src/test/java/example/FelineTest.java
@@ -1,38 +1,42 @@
 package example;
 
 import com.example.Feline;
+import com.example.Lion;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-@RunWith(Parameterized.class)
+@RunWith(MockitoJUnitRunner.class)
 public class FelineTest {
     private int actualKittensCount;
     private int expectedKittensCount;
+
     private Feline feline = new Feline();
 
-    public FelineTest(int kittensCount, int expected) {
+    public FelineTest() {
+
+    }
+
+    /*public FelineTest(int kittensCount, int expected) {
         this.actualKittensCount = kittensCount;
         this.expectedKittensCount = expected;
     }
-
+*/
     @Before
     public void setUp() {
         MockitoAnnotations.openMocks(this);
     }
 
-    @Parameterized.Parameters
-    public static Object[] getKittensCount() {
-        return new Object[][]{{0, 0}, {1, 1}, {10, 10}};
-    }
 
     @Test
     public void eatMeatTest() throws Exception {
@@ -47,18 +51,13 @@ public class FelineTest {
     }
 
     @Test
-    public void getKittensParametrizedTest() {
-        assertEquals(expectedKittensCount, feline.getKittens(actualKittensCount));
-    }
-
-    @Test
     public void getKittensNotParametrizedTest() {
-        feline.getKittens();
         assertEquals(expectedKittensCount, feline.getKittens(actualKittensCount));
     }
 
     @Test
-    public void getFoodTest() throws Exception {
+    public void getFoodTest() {
+        Feline feline = Mockito.mock(Feline.class);
         assertThrows(Exception.class, () -> feline.getFood(Mockito.anyString()));
     }
